+++ src/browser/cdp.ts
@@ -421,6 +421,11 @@ class CDPPage extends BasePage {
 
   async nativeClick(x: number, y: number): Promise<void> {
     await this.cdp('Input.dispatchMouseEvent', {
+      type: 'mouseMoved',
+      x,
+      y,
+    });
+    await this.cdp('Input.dispatchMouseEvent', {
       type: 'mousePressed',
       x,
       y,